% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interCross.R
\docType{package}
\name{aaa-interCross}
\alias{aaa-interCross}
\title{Intervention analysis using the cross section of a wide panel}
\description{
\code{interCross} uses the variation of successive observations in the
cross section of a wide panel to identify the law of motion
for conduction an intervention analysis.
\code{interCross} creates a discrete approximation to a low-order Markov
process defined on a continuous state space in discrete time.
Once the state space is discretized, \code{interCross} provides tools
to estimate the transition matrices and analyze the Markov process.
}
\seealso{
\code{multinom} function in \code{nnet} package for estimating the
transition matrices.
\code{DTMCPack} and \code{markovchain} for analyzing the discrete-time Markov
model once the continuous state space is discretized.
}
