% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discCtsDTMC.R
\name{cut_states}
\alias{cut_states}
\title{Discretize a Variable Defined on a Continuous State Space}
\usage{
cut_states(x_cts, breaks)
}
\arguments{
\item{x_cts}{a numeric vector of observations from a continuous random
variable with discontinuities in the distribution function.}

\item{breaks}{a numeric vector of thresholds for allocating elements of the
continuous state space to discrete state space.}
}
\value{
a categorical variable with states for each variable that
correspond to elements in the continuous state space.
}
\description{
\code{cut_states} transforms a variable defined on a continuous state space
into a discrete variable.
}
