% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discCtsDTMC.R
\docType{package}
\name{discretizeCtsDTMC}
\alias{discretizeCtsDTMC}
\title{Discretize a Continuous-valued Discrete-time Markov Chain}
\description{
\code{discretizeCtsDTMC} creates a discrete approximation to a Markov
process defined on a continuous state space in discrete time.
Once the state space is discretized, \code{discretizeCtsDTMC} provides tools
to estimate the transition matrices and analyze the Markov process.
}
\note{
We'll think of a better name later.
}
\seealso{
\code{multinom} function in \code{nnet} package for estimating the
transition matrices.
\code{DTMCPack} and \code{markovchain} for analyzing the discrete-time Markov
model once the continuous state space is discretized.
}
