% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interCross.R
\name{find_atoms}
\alias{find_atoms}
\title{Find Points of Discontinuity in the CDF}
\usage{
find_atoms(x, min_atom = 0.05)
}
\arguments{
\item{x}{a numeric vector of observations from a continuous random
variable with discontinuities in the distribution function.}

\item{min_atom}{a numeric scalar on the unit interval which is the minimum
propability density to detect a discontinuity in the CDF of x.}
}
\value{
a numeric vector of points in the sample space at which there are
discontinuities in the distribution function, which are sometimes referred to
as "atoms".
}
\description{
\code{find_atoms} finds points of discontinuity in the CDF of a
continuous random variable.
}
\examples{
find_atoms(x = c(rep(3,10), seq(1, 100), rep(26, 10)))
find_atoms(x = c(rep(3,10), seq(1, 100), rep(26, 3)))
find_atoms(x = c(rep(3,10), seq(1, 100), rep(26, 3)), min_atom = 0.025)
}
