% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interCross.R
\name{forecast_distn}
\alias{forecast_distn}
\title{Forecast a Probability Distribution}
\usage{
forecast_distn(trans_mats, init_probs, n_ahead)
}
\arguments{
\item{trans_mats}{an array of transition matrices for the Markov process.}

\item{init_probs}{a numeric probability vector that defines the initial
proportions of the population in each state.}

\item{n_ahead}{an integer number of lags that defines the order of the
Markov process.}
}
\value{
a \code{n_ahead}-row numeric matrix of probability vectors that
define the forecasted proportion of the population in each state at each time.
}
\description{
\code{forecast_distn} calculates a forecast of a probability distribution
for a population governed by a discrete-state, discrete-time Markov process.
}
