% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interCross.R
\name{forecast_k_probs}
\alias{forecast_k_probs}
\title{Compute a Time Series of Probability Vectors.}
\usage{
forecast_k_probs(
  start_probs,
  P_hat,
  num_steps,
  date_list = NULL,
  mat_num_list = NULL
)
}
\arguments{
\item{start_probs}{is a probability vector from which the forecasts will satrt.}

\item{date_list}{is a list of dates or time index of integers.}
}
\value{
a matrix with a series of forecasted probability
vectors in the rows, with one row for each time stamp in a data frame.
Each forecast is a k-step-ahead forecast starting from the probability
vector on the first date for k from 1 to \code{num_steps}.
}
\description{
\code{forecast_k_probs} calculates a matrix with a series of probability
vectors in the rows, with one rows for each time stamp in a data frame.
}
