% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interCross.R
\name{KLD_forecast_test}
\alias{KLD_forecast_test}
\title{Kullback-Leibler Tests for Divergence between Two Sets of Distributions}
\usage{
KLD_forecast_test(forecast_probs, observed_probs, num_obs)
}
\arguments{
\item{forecast_probs}{a matrix with a series of forecasted probability
vectors in the rows, with one row for each time stamp in a data frame.}

\item{observed_probs}{a matrix with a series of probability
vectors in the rows, with one rows for each time stamp in a data frame.}

\item{num_obs}{a scalar or vector integer number of sample sizes,
If it is a scalar, it is assumed that the pane is balanced, otherwise,
a vector of sample sizes should be provided for each time stamp.}
}
\value{
a data frame of KLD statistics and p-values.
}
\description{
\code{KLD_step} calculates the Kullback-Leibler divergence statistic
to forecast accuracy of k-step-ahead forecasts.
}
