% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interCross.R
\name{KLD_gof_1step}
\alias{KLD_gof_1step}
\title{Goodness of Fit Tests in Sample}
\usage{
KLD_gof_1step(dt, P_hat, date_list, mat_num_list)
}
\arguments{
\item{dt}{is a \code{data.table}}

\item{P_hat}{an array or matrix of transition matrices.}

\item{date_list}{is a list of dates or time index of integers.}

\item{mat_num_list}{is a list of integers representing the index number of
the third dimension of \code{P_hat}, if \code{P_hat} is an array,
for selecting the relevant transition matrix for each time in \code{date_list}.}
}
\value{
a data frame of KLD statistics and p-values.
}
\description{
\code{KLD_gof_1step} calculates the Kullback-Leibler divergence statistic
to measure the goodness of fit of the 1-step-ahead forecasts with a set of
transition matrices.
}
\note{
: Functionality for different date types not implemented.
}
