% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interCross.R
\name{prob_series}
\alias{prob_series}
\title{Compute a Time Series of Probability Vectors.}
\usage{
prob_series(dt, date_list = NULL)
}
\arguments{
\item{dt}{is a \code{data.table}}

\item{date_list}{is a list of dates or time index of integers.}
}
\value{
a matrix with a series of probability
vectors in the rows, with one rows for each time stamp in a data frame.
}
\description{
\code{prob_series} calculates a matrix with a series of probability
vectors in the rows, with one rows for each time stamp in a data frame.
}
