% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interCross.R
\name{reshape_markov_data}
\alias{reshape_markov_data}
\title{Reshape Data for Estimation of Transition Matrices.}
\usage{
reshape_markov_data(dt, time_unit, n_lags)
}
\arguments{
\item{dt}{a \code{data.table} object with specific columns
\code{x_disc}, \code{id} and \code{time}.
It must be sorted by \code{id} then \code{time} to ensure the
consecutive listing of observations for each individual.}

\item{time_unit}{a string to identify the unit of time between consecutive
observations. Can be "month" or "integer".}

\item{n_lags}{an integer number of lags to define the order of the Markov process.
for each column of the transition matrices (default is \code{FALSE}).}

\item{x_disc}{a numeric vector of observations from a discrete random variable.
If x is a matrix, it is assumed that the first two columns are the id and the time stamp.}

\item{time}{a vector of time stamps that correspond to the observations in \code{x}.}

\item{id}{a vector of labels to identify different individuals in the population.}
}
\value{
x_mat a matrix of data for estimation of transition matrices.
It will have 4 columns: the id, the time stamp, the observations and a boolean
variable to indicate valid observations for estimation.
It will be reordered, first by id, then time.
}
\description{
\code{reshape_markov_data} reshapes data from a discrete-state,
discrete-time Markov process and organizes it into a form suitable to
estimate transition matrices .
}
