% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discCtsDTMC.R
\name{sim_pop_MC}
\alias{sim_pop_MC}
\title{Simulate a Population of Individuals following a Markov Process}
\usage{
sim_pop_MC(trans_mats, init_probs, n_ahead, n_ind)
}
\arguments{
\item{trans_mats}{an array of transition matrices for the Markov process.}

\item{init_probs}{a numeric probability vector that defines the initial
proportions of the population in each state.}

\item{n_ahead}{an integer number of lags that defines the order of the
Markov process.}

\item{n_ind}{an integer number of individuals in the cross-section.}
}
\value{
a data frame of realizations for individuals in the cross-section,
including id labels for individuals and time stamps for each time period.
}
\description{
\code{sim_pop_MC} draws a realization of a population of individuals,
each following a discrete-time Markov process. It can be used for a
parametric bootstrap procedure when the large sample properties may not hold.
}
