% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interCross.R
\name{state_breaks}
\alias{state_breaks}
\title{Divide a Continuous State Space into Intervals}
\usage{
state_breaks(x, num_breaks)
}
\arguments{
\item{x}{a numeric vector of observations from a continuous random
variable with discontinuities in the distribution function.}

\item{num_breaks}{is an integer that captures the number of break points
to use when dividing the support of \code{x} into intervals.}
}
\value{
a numeric vector of thresholds for allocating elements of the
continuous state space to a discrete state space.
}
\description{
\code{state_breaks} calculates a vector of thresholds for allocating
elements of a continuous state space to a discrete state space
}
\examples{
state_breaks(x = seq(-5,10), num_breaks = 4)

}
