% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discCtsDTMC.R
\name{test_fore_dev}
\alias{test_fore_dev}
\title{Test for a Deviation from Forecasted Population}
\usage{
test_fore_dev(x, time_stamp, id, out_probs)
}
\arguments{
\item{x}{a numeric vector of observations of a discrete random variable
from the out-of-sample forecast period.}

\item{time_stamp}{a vector of time stamps that correspond to the observations in \code{x}.}

\item{id}{a vector of labels to identify different individuals in the population.}

\item{out_probs}{a \code{n_ahead}-row numeric matrix probability vector that
defines the forecasted proportions of the population in each state.}
}
\value{
a data frame of Kullback-Leibler divergence statistic
and p-values from the quantiles of the chi-squared distribution.
}
\description{
\code{test_fore_dev} does something.
}
