% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discCtsDTMC.R
\name{find_atoms}
\alias{find_atoms}
\title{Find Points of Discontinuity in the CDF}
\usage{
find_atoms(x)
}
\arguments{
\item{x}{a numeric vector of observations from a continuous random
variable with discontinuities in the distribution function.}
}
\value{
a numeric vector of points in the sample space at which there are
discontinuities in the distribution function, which are sometimes referred to
as "atoms".
}
\description{
\code{find_atoms} finds points of discontinuity in the CDF of a
continuous random variable.
}
