% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discCtsDTMC.R
\name{state_breaks}
\alias{state_breaks}
\title{Divide a Continuous State Space into Intervals}
\usage{
state_breaks(x)
}
\arguments{
\item{x}{a numeric vector of observations from a continuous random
variable with discontinuities in the distribution function.}
}
\value{
a numeric vector of thresholds for allocating elements of the
continuous state space to discrete state space.
}
\description{
\code{state_breaks}
}
